Fix handling of backslash escapes in CSS names
We should not include these backslashes in the generated Rust code, because
that leads to HTML with names like `w-0\.5` (with a literal backslash). These
names don't need to be escaped in HTML, only in the CSS file that Tailwind
generates.

This also required a fix to the tailwind extractor JS code.